v1.0.0

+++ routers/refresh.go
@@ -41,7 +41,7 @@ func (this *RefreshRouter) Get() {
 	}
 
 	pdoc, err := doc.CheckDoc(q, "", doc.RT_Refresh)
-	if err == nil {
+	if err == nil && pdoc != nil {
 		os.Remove("." + utils.DocsJsPath + pdoc.ImportPath + ".js")
 
 		this.Redirect("/"+q, 302)
